+++ src/main/java/com/etiya/academy/repository/ProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long>
+public interface ProductRepository extends JpaRepository<Product, Integer>
 {
     List<Product> findAll();
 //    Product save(Product product);
